Merge from upstream

+++ packaging/macosx/make-package.sh
@@ -16,35 +16,39 @@ DMGCREATE=${DIR}/yoursway-create-dmg/create-dmg
 VERSION=$(cd ${DIR}/subsurface; ./scripts/get-version linux)
 
 # first build and install Subsurface and then clean up the staging area
+# make sure we didn't lose the minimum OS version
 rm -rf ./Subsurface.app
+cmake -DCMAKE_OSX_DEPLOYMENT_TARGET=10.11 -DCMAKE_OSX_SYSROOT=/Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX10.11.sdk .
 LIBRARY_PATH=${DIR}/install-root/lib make -j8
 LIBRARY_PATH=${DIR}/install-root/lib make install
 
 # now adjust a few references that macdeployqt appears to miss
 EXECUTABLE=Subsurface.app/Contents/MacOS/Subsurface
-for i in libssh libssrfmarblewidget libgit2 libGrantlee_TextDocument.dylib libGrantlee_Templates.dylib; do
+for i in libgit2 libGrantlee_TextDocument.dylib libGrantlee_Templates.dylib; do
 	OLD=$(otool -L ${EXECUTABLE} | grep $i | cut -d\  -f1 | tr -d "\t")
-	if [ ! -z ${OLD} ] ; then
+	if [[ ! -z ${OLD} && ! -f Subsurface.app/Contents/Frameworks/$(basename ${OLD}) ]] ; then
+		# copy the library into the bundle and make sure its id and the reference to it are correct
 		cp ${DIR}/install-root/lib/$(basename ${OLD}) Subsurface.app/Contents/Frameworks
 		SONAME=$(basename $OLD)
 		install_name_tool -change ${OLD} @executable_path/../Frameworks/${SONAME} ${EXECUTABLE}
-		if [[ "$i" = "libssh" ]] ; then
-			LIBSSH=$(basename ${OLD})
-		fi
-		if [[ "$i" = "libgit2" && ! -z ${LIBSSH} ]] ; then
-			CURLIBSSH=$(otool -L Subsurface.app/Contents/Frameworks/${SONAME} | grep libssh | cut -d\  -f1 | tr -d "\t")
-			install_name_tool -change ${CURLIBSSH} @executable_path/../Frameworks/${LIBSSH} Subsurface.app/Contents/Frameworks/${SONAME}
+		install_name_tool -id @executable_path/../Frameworks/${SONAME} Subsurface.app/Contents/Frameworks/${SONAME}
+		# also fix incorrect references inside of libgit2
+		if [[ "$i" = "libgit2" ]] ; then
+			CURLLIB=$(otool -L Subsurface.app/Contents/Frameworks/${SONAME} | grep libcurl | cut -d\  -f1 | tr -d "\t")
+			install_name_tool -change ${CURLLIB} @executable_path/../Frameworks/$(basename ${CURLLIB}) Subsurface.app/Contents/Frameworks/${SONAME}
+			# SSHLIB=$(otool -L Subsurface.app/Contents/Frameworks/${SONAME} | grep libssh2 | cut -d\  -f1 | tr -d "\t")
+			# install_name_tool -change ${SSHLIB} @executable_path/../Frameworks/$(basename ${SSHLIB}) Subsurface.app/Contents/Frameworks/${SONAME}
 		fi
 	fi
 done
+
+# next, copy libssh2.1
+# cp ${DIR}/install-root/lib/libssh2.1.dylib Subsurface.app/Contents/Frameworks
+
+# next, replace @rpath references with @executable_path references in Subsurface
 RPATH=$(otool -L ${EXECUTABLE} | grep rpath  | cut -d\  -f1 | tr -d "\t" | cut -b 8- )
 for i in ${RPATH}; do
 	install_name_tool -change @rpath/$i @executable_path/../Frameworks/$i ${EXECUTABLE}
-done
-MARBLELIB=$(ls Subsurface.app/Contents/Frameworks/libssrfmarblewidget*dylib)
-RPATH=$(otool -L ${MARBLELIB} | grep rpath  | cut -d\  -f1 | tr -d "\t" | cut -b 8- )
-for i in ${RPATH}; do
-	install_name_tool -change @rpath/$i @executable_path/../Frameworks/$i ${MARBLELIB}
 done
 
 # next deal with libGrantlee
@@ -53,22 +57,23 @@ for i in QtScript.framework/Versions/5/QtScript QtCore.framework/Versions/5/QtCo
 	install_name_tool -change @rpath/$i @executable_path/../Frameworks/$i ${LIBG}
 done
 
-# it seems the compiler in XCode 4.6 doesn't build Grantlee5 correctly,
-# so cheat and copy over pre-compiled binaries created with a newer compiler
-# and adjust their references to the Grantlee template library
-#
-# -disabled for now as this is still under more investigation-
-# cp -a /Users/hohndel/src/tmp/Subsurface.app/Contents Subsurface.app/
-cp ${DIR}/tmp/Subsurface.app/Contents/Frameworks/lib{sql,usb,zip}* Subsurface.app/Contents/Frameworks
-
 # clean up shared library dependency in the Grantlee plugins
 for i in Subsurface.app/Contents/PlugIns/grantlee/5.0/*.so; do
 	OLD=$(otool -L $i | grep libGrantlee_Templates | cut -d\  -f1 | tr -d "\t")
 	SONAME=$(basename $OLD )
 	install_name_tool -change ${OLD} @executable_path/../Frameworks/${SONAME} $i;
+	OLD=$(otool -L $i | grep QtCore | cut -d\  -f1 | tr -d "\t")
+	install_name_tool -change ${OLD} @executable_path/../Frameworks/QtCore.framework/QtCore $i;
 	mv $i Subsurface.app/Contents/PlugIns/grantlee
 done
 rmdir Subsurface.app/Contents/PlugIns/grantlee/5.0
+pushd Subsurface.app/Contents/PlugIns/grantlee
+ln -s . 5.0
+popd
+
+if [ "$1" = "-nodmg" ] ; then
+	exit 0
+fi
 
 # copy things into staging so we can create a nice DMG
 rm -rf ./staging
